last version before fixing the exel

+++ JobCentre/Views/DataBaseForm.cs
@@ -15,19 +15,31 @@ namespace JobCentre.Views
 {
     public partial class DataBaseForm : Form
     {
-        string primaryKey;
         string connectionString = Helper.connectionString;
-        public string temp; // change in future
 
+        string primaryKey
+        {
+            get
+            {
+                if (tableName != "Vacancy request")
+                    return MainDataGridView.Columns[0].Name;
+                return "Vacancy request";
+            }
+        }
         public string sqlString = "";
         public DataTable filterDataTable;
         string tableName { get { return tabControl.SelectedTab.AccessibleDescription; } }
-
+        List<string> headers;
         public DataBaseForm()
         {
             InitializeComponent();
             tabControl_SelectedIndexChanged(null, null);
+
+            ToolTip tp = new ToolTip();
+            tp.SetToolTip(EditButton, "If you choose more than 1 item, you will edit only first");
+            tp.SetToolTip(DeleteButton, "If you choose more than 1 item, you will delete only first");
         }
+
 
         private void DataBaseForm_Load(object sender, EventArgs e)
         {
@@ -44,7 +56,7 @@ namespace JobCentre.Views
                 int rowIndex = MainDataGridView.CurrentRow.Index;
                 return (int)MainDataGridView.Rows[rowIndex].Cells[0].Value;
             }
-        } //  с индексами надо разобраться это не все
+        }
 
         int secondColumn
         {
@@ -104,12 +116,29 @@ namespace JobCentre.Views
         private void DeleteButton_Click(object sender, EventArgs e)
         {
             if (selectedIndex == -1)
+            {
                 return;
+            }
+            SureForm form = new SureForm();
+            form.ShowDialog();
+            if (form.DialogResult != DialogResult.Yes)
+            {
+                return;
+            }
+
             using (SqlConnection sql = new SqlConnection(connectionString))
             {
-                string sqlCopy = (string)sqlString.Clone();
-                sqlCopy = String.Format("DELETE FROM [dbo].[{0}] WHERE [{1}] = {2};", tableName, primaryKey, selectedIndex);
-                SqlDataAdapter adapter = new SqlDataAdapter(sqlCopy, sql);
+                if (primaryKey != "Vacancy request")
+                    sqlString = String.Format("DELETE FROM [dbo].[{0}] WHERE [{1}] = {2};", tableName, primaryKey, selectedIndex);
+                else
+                {
+                    // no need any try-catch, because this table isn't linked to anyone
+                    vacancy_requestTableAdapter.DeleteVacRequestQuery(selectedIndex, secondColumn, thirdColumn);
+                    vacancy_requestTableAdapter.Update(this.laborExchangeDataSet);
+                    tabControl_SelectedIndexChanged(null, null);
+                    return;
+                }
+                SqlDataAdapter adapter = new SqlDataAdapter(sqlString, sql);
                 DataTable dt = new DataTable();
                 try
                 {
@@ -122,35 +151,11 @@ namespace JobCentre.Views
             }
             employeeTableAdapter.Update(this.laborExchangeDataSet);
             employerTableAdapter.Update(this.laborExchangeDataSet);
-            vacancy_requestTableAdapter.Update(this.laborExchangeDataSet);
+            //vacancy_requestTableAdapter.Update(this.laborExchangeDataSet);
             vacancyTableAdapter.Update(this.laborExchangeDataSet);
             tabControl_SelectedIndexChanged(null, null);
         }
 
-        private void SearchTextBox_TextChanged(object sender, EventArgs e)
-        {
-            if (SearchTextBox.Text == "")
-            {
-                for (int i = 0; i < MainDataGridView.Rows.Count - 1; i++) // последняя строка пустая для добавления
-                    for (int j = 0; j < MainDataGridView.Columns.Count; j++)
-                    {
-                        MainDataGridView.Rows[i].Cells[j].Selected = false;
-                    }
-                return;
-            }
-            string searchText = SearchTextBox.Text;
-            for (int i = 0; i < MainDataGridView.Rows.Count - 1; i++)
-                for (int j = 0; j < MainDataGridView.Columns.Count; j++)
-                {
-                    MainDataGridView.Rows[i].Cells[j].Selected = false;
-                }
-            for (int i = 0; i < MainDataGridView.Rows.Count - 1; i++)
-                for (int j = 0; j < MainDataGridView.Columns.Count; j++)
-                {
-                    if (MainDataGridView.Rows[i].Cells[j].Value.ToString().ToLower().Contains(searchText.ToLower()))
-                        MainDataGridView.Rows[i].Cells[j].Selected = true;
-                }
-        }
 
         private void FilterButton_Click(object sender, EventArgs e)
         {
@@ -169,12 +174,16 @@ namespace JobCentre.Views
                         return;
                     break;
                 case "Vacancy request":
-                    VacancyFilterForm vff = new VacancyFilterForm();
+                    VacancyRequestFilterForm vrff = new VacancyRequestFilterForm(this);
+                    vrff.ShowDialog();
+                    if (vrff.DialogResult != DialogResult.OK)
+                        return; 
+                        break;
+                case "Vacancy":
+                    VacancyFilterForm vff = new VacancyFilterForm(this);
                     vff.ShowDialog();
                     if (vff.DialogResult != DialogResult.OK)
                         return;
-                    break;
-                case "Vacancy":
                     break;
             }
             MainDataGridView.DataSource = filterDataTable;
@@ -183,7 +192,6 @@ namespace JobCentre.Views
         private void Clear_button_Click(object sender, EventArgs e)
         {
             SearchTextBox.Text = "";
-            temp = "";
             sqlString = "";
             using (SqlConnection sql = new SqlConnection(connectionString))
             {
@@ -195,33 +203,46 @@ namespace JobCentre.Views
             }
         }
 
-        private void tabControl_SelectedIndexChanged(object sender, EventArgs e)
+        public void tabControl_SelectedIndexChanged(object sender, EventArgs e)
         {
-            switch (tableName) // селектед айди неверный
-            {
-                case "Employer":
-                    //FindListComboBox.DataSource = new List<string> { "Employer's ID", "Full name", "Pasport number", "Phone number", "Bank details", 
-                    //"Individual tax number", "Main property type", "Economical activity"};
-                    primaryKey = "Employer's ID";
-                    break;
-                case "Employee":
-                    //FindListComboBox.DataSource = new List<string> { "Employee's ID", "Full name", "Profession", 
-                    //    "Record of service", "Special skills", "Location", "Identificaton code", 
-                    //"Education", "Needs", "Work mode", "Work nature"};
-                    primaryKey = "Employee's ID";
-                    break;
-                case "Vacancy request":
-                    //FindListComboBox.DataSource = new List<string> { "Vacancy ID", "Employee's ID", "Submission date", "Interview date", "Result" };
-                    primaryKey = "Vacancy ID";
-                    break;
-                case "Vacancy":
-                    //FindListComboBox.DataSource = new List<string> { "Vacancy Id", "Work location", "Social package","Scope",  "Beginning of work", "End of work",
-                    //"Salary", "Features", "Possibility of working minors", "Position", "Type of ownership", "Company name"};
-                    primaryKey = "Vacancy Id";
-                    break;
-            }
             Clear_button_Click(null, null);
             filterDataTable = null;
+            headers = new List<string>();
+            foreach (DataGridViewColumn i in MainDataGridView.Columns)
+                headers.Add(i.Name);
+            currentHeaderComboBox.DataSource = headers;
+
         }
+
+        private void SearchTextBox_TextChanged(object sender, EventArgs e)
+        {
+            for (int i = 0; i < MainDataGridView.Rows.Count - 1; i++)
+                for (int j = 0; j < MainDataGridView.Columns.Count; j++)
+                {
+                    MainDataGridView.Rows[i].Cells[j].Selected = false;
+                }
+            if (SearchTextBox.Text == "")
+            {
+                return;
+            }
+            string searchText = SearchTextBox.Text;
+            string currentHeader = currentHeaderComboBox.Text;
+            int index = headers.IndexOf(currentHeader);
+            for (int i = 0; i < MainDataGridView.Rows.Count - 1; i++)
+            {
+                if (MainDataGridView.Rows[i].Cells[index].Value.ToString().ToLower().Contains(searchText.ToLower()))
+                    MainDataGridView.Rows[i].Cells[index].Selected = true;
+            }
+        }
+
+        private void currentHeaderComboBox_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            for (int i = 0; i < MainDataGridView.Rows.Count - 1; i++)
+                for (int j = 0; j < MainDataGridView.Columns.Count; j++)
+                {
+                    MainDataGridView.Rows[i].Cells[j].Selected = false;
+                }
+        }
+
     }
 }
